1.1.Create those Subscription Plans
Yes! You're back! You survived Stripe part 1. Ok... that's cool, but now we have bigger fish to fry. So fasten your seat belts, put on a pot of coffee, and get ready to sharpen your Stripe subscription skills.

Tell Stripe about the Plans

Our first step is to tell Stripe about these two plans. To do that, open your trusty Stripe dashboard, make sure you're in the test environment, and under "Billing" find "Products" on the left. We're going to create two products and attach to each of them one plan in Stripe by hand.
Click "New" button. Set "Product name" to farmer_brent. The other two inputs are optional. Click "Create product". Now we need to add a plan to this product in this window. Set "Plan nickname" to farner_brent_monthly and let stripe to give it an id. set the price as $99. By default the customer will be billed monthly so will let the rest of the inputs as they are. Repeat the same steps for new_zeelander product and attach the new_zeelander monthly plan to it but for $199.
And yes, later, we'll add yearly versions of each plan... but one step at a time!