1.2.Create those Subscription Plans(Manage the Plans in Our Code)
Manage the Plans in Our Code
Now, head back to our code and open the src/Subscription directory. I added two classes to help us stay organized. The first is called SubscriptionHelper.
And as you can see... it's not very helpful... yet. But pretty soon, it will keep track of the two plans we just added. We'll use the second class - SubscriptionPlan.
To hold the data for each plan: the plan ID, name and price. But we won't save these to the database.
In SubscriptionHelper, setup the two plans. For the first, use id provided by stripe for the key, name it farmer_brent_monthly and use 99 as the pricing.
Copy that and repeat the same thing: id provided by stripe, new_zealander_monthly and 199.
Love it! The 2 plans live in Stripe and in our code. Time to make it possible to add these to our cart, and then checkout.