2.1.Add the Subscription to your Cart
We can already add products to our cart... but a user should also be able to click these fancy buttons and add a subscription to their cart.
Open up OrderController: the home for the checkout and shopping cart magic. I've already started a new page called addSubscriptionToCartAction().
When we're done, if the user goes to /cart/subscription/farmer_brent_monthly, this should put that into the cart.
First, hook up the buttons to point here. The template for this page lives at templates/views/product/pricing.html.twig.
I started adding the link to this page, but left the plan ID blank. Fill 'em in!
Go back to that page and refresh. The links look great