2.2.Add the Subscription to your Cart{Put the Plan in the Cart}
Put the Plan in the Cart
Now back to the controller! In the first Stripe tutorial, we worked with a ShoppingCart class that I created for us... because it's not really that important. It basically allows you to store products and a subscription in the user's session, so that as they surf around, we know what they have in their cart.

Next, add $plan = ->subscriptionHelper->findPlan() and pass it the $planId. So this is nice: we give it the plan ID, and it gives us the corresponding, wonderful, SubscriptionPlan object. But if the $planId doesn't exist for some reason, throw $this->createNotFoundException() to cause a 404 page. Finally, add the plan to the cart, with $this->cart->addSubscription() and pass it the plan ID.
And boom! Our cart knows about the subscription! Finally, send them to the checkout page with return $this->redirectToRoute('order_checkout') - that's the name of our checkoutAction route
Okay team, give it a try! Add the Farmer Brent plan. Bah! We need to login: use the pre-filled email and the password used by all sheep: breakingbaad.

Ok, this looks kinda right: the total is 9 because the ShoppingCart object knows about the subscription... but we haven't printed anything about the subscription in the cart table. So it looks weird.