4.3.Give the User a Subscription (in our Database){Prepping the Account Page}
Prepping the Account Page
And that's cool because we'll be able to fill in this fancy account page I created! Open up the template for this page at templates/views/profile/account.html.twig. Instead of "None", add an if statement: if app.user - that's the currently-logged-in user app.user.subscription, then we know they have a Subscription. Add a label that says "Active". If they don't have a subscription, say "None".
If you refresh now... it says None. We actually do have a Subscription in Stripe from a moment ago, but our database doesn't know about it. That's what we need to fix.