4.4.Give the User a Subscription (in our Database){Updating the Database}
Updating the Database
Since our goal is to update the database during checkout, go back to OrderController and find the chargeCustomer() method that holds all the magic.
But instead of putting the code to update the database right here, let's add it to SubscriptionHelper: this class will do all the work related to subscriptions. Add a new method at the bottom called public function addSubscriptionToUser() with two arguments: the \Stripe\Subscription object that was just created and the User that the Subscription should belong to.
Inside, start with $subscription = $user->getSubscription(). So, the user may already have a row in the subscription table from a previous, expired subscription. If they do, we'll just update that row instead of creating a second row. Every User will have a maximum of one related row in the subscription table. It keeps things simple.
But if they don't have a previous subscription, let's create one: $subscription = new Subscription(). Then, $subscription->setUser($user)
Our other todo is to update the fields on the Subscription object: $stripeSubscriptionId and $stripePlanId. To keep things clean, open Subscription and add a new method at the bottom: public function activateSubscription() with two arguments: the $stripePlanId and $stripeSubscriptionId.
Set each of these onto the corresponding properties. Also add $this->endsAt = null.
We'll talk more about that later, but this field will help us know whether or not a subscription has been cancelled.
Back in SubscriptionHelper, call $subscription->activateSubscription()
We need to pass this the stripePlanId and the stripeSubscriptionId. But remember! We have this fancy \Stripe\Subscription object! In the API docs, you can see its fields, like id and plan with its own id sub-property.
Cool! Pass the method $stripeSubscription->plan->id and $stripeSubscription->id.
Booya!
And, time to save this to the database! Since we're using Doctrine in Symfony, we need the EntityManager object to do this. I'll use dependency injection: add an EntityManager argument to the __construct() method, and set it on a new $em property.
For Symfony users, this service is using auto-wiring. So because I type-hinted this with EntityManager, Symfony will automatically know to pass that as an argument.
Finally, at the bottom, add $this->em->persist($subscription) and $this->em->flush($subscription) to save just the Subscription
With all that setup, go back to OrderController to call this method. To do that, we need the \Stripe\Subscription object. Fortunately, the createSubscription method returns this.
So add $stripeSubscription = in front of that line. Then, add $this->subscriptionHelper->addSubscriptionToUser() passing it $stripeSubscription and the currently-logged-in $user
Phew! That may have seemed like a lot, but ultimately, this line just makes sure that there is a subscription row in our table that's associated with this user and up-to-date with the subscription and plan IDs.
Let's go try it out. Add a new subscription to your cart, fill out the fake credit card information and hit checkout. No errors! To the account page! Yes! The subscription is active! Our database is up-to-date.