5.3.Data: Card Last 4 Digits{The Card Details on the Stripe Customer;Setting the Card Details}
The Card Details on the Stripe Customer
Head to the Stripe API docs and click on Customers. The card information is attached to the customer under a field called sources. Yes, sources with an s at the end because you could attach multiple cards to a customer if you wanted. But we're not: on checkout, we set just one card on the customer, and replace any existing card, if there was one.
In other words, sources will always have just one entry. That one entry will have a data key, and that will describe the card: giving us all the info you see here.
Now to the plan: use the Stripe API to populate the card information on the User table right during checkout.

Setting the Card Details
In OrderController::chargeCustomer(), we either create or retrieve the \Stripe\Customer. Assign both calls to a new $stripeCustomer variable. In StripeClient, the createCustomer() method already returns the \Stripe\Customer object, so we're good here.
The updateCustomerCard() method, however, retrieves the customer... but gets lazy and doesn't return it. Fix that with return $customer.
Back in OrderController, we've got the \Stripe\Customer... so we're mega dangerous! But instead of updating the fields on User right here, let's do it in SubscriptionHelper. Add a new public function updateCardDetails() method with a User object that should be updated and the \Stripe\Customer object that's associated with it.
Now, this is pretty easy: $cardDetails = $this->stripeCustomer->sources->data[0]. Then, $user->setCardBrand($cardDetails) - go cheat with the Stripe API - the fields we want are brand and last4. So, $cardDetails->brand. And $user->setCardLast4($cardDetails->last4). Save only the user to the database with the classic $this->em->persist($user) and $this->em->flush($user).
Finally, call that method! $this->subscriptionHelper->updateCardDetails() and pass it $user and $stripeCustomer.
No matter how you checkout, we're going to make sure your card details are updated in our database!
Before we try it out and prove how awesome we are, I want to add one more thing: I want to be able to tell the user when they will be billed next.