8.3.Tracking Cancelations in our Database{Making the Account Template Awesome}
We don't want the "Next Billing at" and credit card information to show up.
The hasActiveSubscription() returns true even if the user already cancelled it. Open User: let's add one more method: public function hasActiveNonCanceledSubscription().
Inside, return $this->hasActiveSubscription() && ! $this->getSubscription()->isCancelled(). Use this method in both places in the Twig template.
Refresh one more time! We got it!
But now that the user can cancel, let's make it possible for them to reactivate the subscription. It's actually an easy win.