9.2.Reactivate/Un-cancel my Subscription! {Expired Subscriptions Cannot be Reactivated;Reactivate in Stripe}
Expired Subscriptions Cannot be Reactivated
Let's get to work in the controller. Like everything, this will have two parts. First, we need to reactivate the subscription in Stripe and second, we need to update our database. For the first part, fetch the trusty StripeClient service object.
Next, open that class. Add a new public function reactivateSubscription(). It will need a User argument whose subscription we should reactivate.
As the Stripe docs mentioned, we can only reactivate a subscription that has not been fully canceled. If today is beyond the period end, then the user will need to create an entirely new subscription. That's why we only show the button in our template during this period.
But just in case, add an "if" statement: if ! $user->hasActiveSubscription(), then we'll throw a new exception with the text:

 Subscriptions can only be reactivated if the subscription has not actually ended.

Nothing should hit that code, but now we'll know if something does.

Reactivate in Stripe
To reactivate the Subscription, we first need to fetch it. In the Stripe API docs, find "Retrieve a Subscription." Every object can be fetched using the same retrieve method. Copy this. Then, add, $subscription = and paste. Replace the subscription ID with $user->getSubscription()->getStripeSubscriptionId()
And remember, if any API call to Stripe fails - like because this is an invalid subscription ID - the library will throw an exception. So we don't need to add extra code to check if that subscription was found.
Finally, reactivate the subscription by setting its plan property equal to the original plan ID, which is $user->getSubscription()->getStripePlanId(). Then, send the details to Stripe with $subscription->save(). And just in case, return the $subscription
Love it! Back in ProfileController, reactivate the subscription with, $this->stripeClient->reactivateSubscription($this->getUser())
And we are done on the Stripe side.