9.3.Reactivate/Un-cancel my Subscription! {Updating our Database}
Updating our Database
The other thing we need to worry about - which turns out to be really easy - is to update our database so that this, once again, looks like an active subscription. It's easy, because we've already done the work for this. Check out SubscriptionHelper: we have a method called addSubscriptionToUser(), which is normally used right after the user originally buys a new subscription.
But we can could also call this after reactivating. In reality, this method simply ensures that the Subscription row in the table is up-to-date with the latest stripePlanId, stripeSubscriptionId, periodEnd and endsAt. These last two are the most important: because they changed when we deactivated the subscription. So by calling activateSubscription().
All of that will be reversed, and the subscription will be alive!
Let's do it! In ProfileController, add a $stripeSubscription = in front of the $stripeClient call. Below that, use $this->subscriptionHelper->addSubscriptionToUser() and pass it $stripeSubscription and the current user.
And that is everything!
Give your user a happy flash message and redirect back to the profile page.
I think we're ready to try this! Go back and refresh the profile. Press reactivate and... our "cancel subscription" button is back, "active" is back, "next billing period" is back and "credit card" is back. In Stripe, the customer's most recent subscription also became active again. Oh man, this is kind of fun to play with: cancel, reactivate, cancel, reactivate. The system is solid.