10.3.Cancelation Edge-Case Bugs{Fully Canceling in the Database}
Fully Canceling in the Database
To do that, first return the $stripeSubscription from the cancelSubscription() method. Then, in ProfileController, add $stripeSubscription = before the cancelSubscription() call.
Finally, we can use the status field to know whether or not the subscription has truly been canceled, or if it's still active until the period end. In other words, if $stripeSubscription->status == 'canceled', then the subscription is done! Else, we're canceling at period end and should just call deactivate()
To handle full cancelation, open up Subscription and add a new public function called cancel(). Here, set $this->endsAt to right now, to guarantee that it will look canceled, and $this->billingPeriodEndsAt = null.
In ProfileController, call it: $subscription->cancel()
And we are done!
Now, testing this is a bit difficult. So let's just make sure we didn't break anything major by hitting cancel. Perfect! And we can reactivate.

And this is why subscriptions are hard.