11.2.The Update Card Form! {Hide/Show the Form}
For now, we don't have any errors. In account.html.twig, we could pass an error variable to the include() call. But, we could also do it in ProfileController::accountAction(). Add error set to null.
Refresh and click "Update Card". We are in business!
But, this has two cosmetic problems. First, the button says "Checkout" ! That's a little scary, and misleading. Let's change it!