11.3.The Update Card Form! {Fixing the Button Text}
Fixing the Button Text
In _cardForm.html.twig, replace "Checkout" with a new variable called buttonText|default('Checkout').
So, if the variable is not defined, it'll print "Checkout".
Now, override that in account.html.twig. Give include() a second argument: an array of extra variables. Pass buttonText as Update Card.
Refresh! Cool, button problem solved.
The second problem is pretty obvious if you fill out the card number field on checkout, and then compare it with the profile page! On the checkout page, we included a lot of JavaScript that does cool stuff like format this field. And, much more importantly, the JS is also responsible for sending the credit card information to Stripe, fetching the token, putting it in the form, and submitting it. We definitely still need that.