11.4.The Update Card Form! {Sharing Card JavaScript}
Sharing Card JavaScript
Ok, how can we reuse the JavaScript? In checkout.html.twig, we just inlined all of our JS right in the template. That's not great, but since this isn't a course about JavaScript, let's solve this as easily as possible. Copy all of the JS and create a new template called _creditCardFormJavaScript.html.twig inside the order/ directory. Paste this there.
Now, in checkout, include that template! Copy that and include the same thing in account.html.twig at the top of the javascripts block.
Ok, refresh and hope for the best! Ah, another missing variable: stripe_public_key:

  Variable "stripe_public_key" does not exist in _creditCardFormJavaScript.html.twig at line 5

We're printing this in the middle of our JS. And the variable comes from OrderController. Copy that line, open ProfileController, and paste it there.
Now the page works! And - at the very least - the JS formatting is rocking.
Frontend stuff is done: let's submit this and update the user's card in Stripe.