12.1.Saving the Updated Card Details{Submitting the Form to the new Endpoint}
When the user fills out this form, our JavaScript sends that info to Stripe, Stripe then sends back a token, we add the token as a hidden field in the form and then submit it. Both the checkout form and update card form will work like this. But, we need to submit the update card form to a new endpoint, that'll update the card, but not charge the user.

Submitting the Form to the new Endpoint
Open ProfileController and add a new endpoint: public function updateCreditCardAction(). Give it the URL /profile/card/update and a fancy name: account_update_credit_card. Add the methods={POST} to be extra cool.
With this in place, we need to update the form to submit here. Check out _cardForm.html.twig. Hmm, the action attribute is empty.
That's because we want the checkout form to submit right back to /checkout. But this won't work for the update card form: it should submit to a different URL.
Instead, render a new variable called formAction and pipe that to the default filter with empty quotes. Now we can override this! In account.html.twig, add another variable to the include: formAction set to path() and the new route name.
Refresh and check out the source. Ok, the form action is ready!