12.2.Saving the Updated Card Details{Saving the Credit Card}
Saving the Credit Card
Let's get to work in ProfileController! But actually... this will be very similar to our checkout logic, so let's go steal code! Copy the line that fetches the stripeToken POST parameter and then head back to ProfileController. Make sure you have a Request argument and the Request use statement.
Then, first, paste that line to fetch the token. Second, fetch the current user object with $user = $this->getUser(). And third, we need to make an API request to stripe that updates the Customer and attaches the token as their new card. That means we'll be using the StripeClient. Fetch it first with $stripeClient = ->stripeClient'
Here's the awesome part: open StripeClient. We already have a method called updateCustomerCard()
We pass the User object and the submitted payment token and it sets it on the Customer and saves.
Victory for code organization! In the controller, just say $this->stripeClient->updateCustomerCard() and pass it $user and $token.
That takes care of things inside Stripe.