16.1.Webhook: Subscription Canceled
Eventually, this function will handle several event types. To handle each, create a switch-case statement: switch ($stripeEvent->type)
That's the field that'll hold one of those many event types we saw earlier.
The first type we'll handle is customer.subscription.deleted. We'll fill in the logic here in a second. Add the break.
We shouldn't receive any other event types because of how we configured the webhook, but just in case, throw an Exception: "Unexpected webhook from Stripe" and pass the type
At the bottom, well, we can return anything back to Stripe. How about a nice message: "Event Handled" and then the type. Well, there is one important piece: you must return a 200-level status code. If you return a non 200 status code, Stripe will think the webhook failed and will try to send it again, over and over again. But 200 means:

  Yo Stripe, it's cool - I heard you, I handled it.