17.3.Testing Webhooks
Go back to the code and find a tutorials directory. This is a special directory I created and you should have it if you downloaded the start code from this page. It has a few things to make our life easier.
Copy the WebhookControllerTest.php file and put it into the tests/ directory. Let's check this out.
This is the start of a test that's specifically written for Symfony. If you're not using Symfony, the code will look different, but the idea is fundamentally the same.
Calling static::createClient() will bootstrap our container in phpunit. Otherwise static::$container will be empty so we would not be able to retrieve services from symfony container.