17.4.Testing Webhooks{Testing Strategy}
Testing Strategy
This test boots Symfony's kernel so that we have access to its container, and all the useful objects inside, like the entity manager! I also added a private function called createSubscription(). We're not using this yet, but by calling it, it will create a new user in the database, give that user an active subscription, and save everything. This won't be a real subscription in Stripe - it'll just live in our database, and will have a random stripeSubscriptionId.

Because here's the strategy:

Create a fake User and fake Subscription in the database;
Send a webhook to /webhooks/stripe with fake JSON, where the subscription id in the JSON matches the fake data in the database;
Verify that the subscription is fully canceled after the webhook finishes.
Add the test: public function testStripeCustomerSubscriptionDeleted()
OK, step 1: create the subscription in the database: $subscription = $this->createSubscription(). Easy! Step 2: send the webhook... which I'll put as a TODO for now. Then step 3: $this->assertFalse() that $subscription->isActive()
Make sense?