17.5.Testing Webhooks{Prepping some Fake JSON}
Prepping some Fake JSON
To send the webhook, we first need to prepare a JSON string that matches what Stripe sends. At the bottom of the class, create a new private function called getCustomerSubscriptionDeletedEvent() with a  argument.
To fill this in, go copy the real JSON from the test webhook. Paste it here with $json = <<<EOF enter, and paste! Now here's the important part: our controller reads the data.object.id key to find the subscription id. Replace this with %s. Then, finish the function with return sprintf($json, $subscriptionId).
Now, this function will create a realistic-looking JSON string, but with whatever subscriptionId we want!
Back in the test function, add $eventJson = $this->getCustomerSubscriptionDeletedEvent() and pass it $subscription->getStripeSubscriptionId(), which is some fake, random value that the function below created