17.6.Testing Webhooks{Sending the Fake Webhook}
In your .env.test file write the lines
STRIPE_PUBLIC_KEY=your public key
STRIPE_SECRET_KEY=your secret key
Your env.local setup is not loaded in test environment so you must let them be known in your tests.

Sending the Fake Webhook
To send the request, create a $client variable set to $this->createClient(). This is Symfony's internal HTTP client: its job is to make requests to our app. If you want, you can also use something different, like Guzzle. It doesn't really matter because - one way or another - you just need to make an HTTP request to the endpoint.
Now for the magic: call $client->request() and pass it a bunch of arguments: POST for the HTTP method, /webhooks/stripe, then a few empty arrays for parameters, files and server. Finally, for the $content argument - the body of the request - pass it $eventJson.
And because things almost never work for me on the first try... and because I know this won't work yet, let's dd($client->getResponse()->getContent()) to see what happened in case there's an error. Also add a sanity check, $this->assertEquals() that 200 matches $client->getResponse()->getStatusCode().
Let's run the test! But not in this video... this video is getting too long. So go get some more coffee and then come back. Then, to the test!