18.1.Testing Part 2: Faking the Event Lookup
Let's run the test. Copy its method name, then open your terminal. It looks like PHPUnit installed just fine. So, run:
php bin/phpunit --filter testStripeCustomerSubscriptionDeleted

 error! Scroll to the top! The webhook returned a 500 error. And if you look closely at the dumped response HTML, you can see the reason:

  No such event: evt_00000000000000

Ah, the id of the fake event that we're sending is evt_00000000000000. That's not a real event in Stripe, and so when the WebhookController reads this and uses Stripe's API to fetch this event, it's not there

61 lines  src/Controller/WebhookController.php
class WebhookController extends AbstractController
{
    public function stripeWebhookAction(Request $request)
    {
        $stripeEvent = $this->stripeClient->findEvent($eventId);
    }
}
It's kind of funny: we added this API lookup to prevent a third-party from sending fake events... and now it's stopping us from doing exactly that. Dang!