18.2.Testing Part 2: Faking the Event Lookup{Faking things in the Test Environment}
Faking things in the Test Environment
Hmm, how to fix this? In the real world, we do want to use stripe's API to fetch the Event object. But in the test environment, this would all work if our code would simply use the JSON we're sending it as the event, and skip the lookup.
Let's do it! We'll set a special configuration variable in the test environment only, then use that to change our logic in the controller.
Open .env.local and add a new parameter: verify_stripe_event set to true. Copy that, and open env_test. Add a parameters key, paste this parameter, but override it to be false.
Now, in WebhookController, we just need an if statement: if $this->getParameter('verify_stripe_event') is true, then keep the normal behavior. Otherwise, set $stripeEvent to json_decode($request->getContent()).
OK, this is not technically perfect: the first $stripeEvent is a \Stripe\Event object, and the second will be an instance of stdClass. But, since you fetch data off both the same way, it should work.

Let's see if does! Try the test again:

php bin/phpunit --filter testStripeCustomerSubscriptionDeleted
This time, no errors! And the dumped response content looks perfect: event handled.

But, the test didn't pass:

 Failed asserting that true is false on line 39

It looks like our webhook is not working, because the subscription is still active. But actually, that's not true: Doctrine is tricking us! In reality, the database has been updated to show that the Subscription is canceled, but this Subscription object is out-of-date.