18.3.Testing Part 2: Faking the Event Lookup{Refreshing the Data after the Test}
Refreshing the Data after the Test
 Query for a fresh one with $subscription = $this->em - I set the EntityManager on that property in setup() - then ->getRepository('App:Subscription') with find($subscription->getId()).
This subscription will have fresh data.

Try the test!

php bin/phpunit --filter testStripeCustomerSubscriptionDeleted
And we are green!

I know, that was kind of hard! But if you want to have automated webhook tests, this is the way to do it. To make matters worse, for other webhooks, you may need to fake additional API calls that you're making to Stripe.

But, there are also a couple of other, manual, but easy ways to test. Let's check 'em out!