22.1.Webhooks: Preventing Replay Attacks
There's one last teeny, tiny little detail we need to worry about with webhooks: replay attacks. These are a security concern but also a practical one.
We already know that nobody can send us, random, fake event data because we fetch a fresh event from Stripe.
But, someone could intercept a real webhook, and send it to us multiple times. I don't know why they would do that, but weird things would happen.
And there's also the practical concern. Suppose Stripe sends us a webhook and we process it. But somehow, there was a connection problem between our server and Stripe, so Stripe never received our 200 status code. Then, thinking that the webhook failed, Stripe tries to send the webhook again. If this were for an invoice.payment_succeeded event, one user might get two subscription renewal emails. That's weird.

Creating the stripe_event_log Table
Let's prevent that. And it's simple: create a database table that records all the event ID's we've handled. Then, query that table before processing a webhook to make sure we haven't seen it before.
In the src/Entity directory, create a new PHP Class called StripeEventLog.
Give it a few properties: $id, $stripeEventId and a $handledAt date field.
Since this project uses Doctrine, I'll add a special use statement on top and then add some annotations, so that this new class will become a new table in the database. Use the "Code"->"Generate" menu, or Command + N on a Mac and select "ORM Class". Repeat that and select "ORM Annotations". Choose all the fields. Update stripeEventId to be a string field - that'll translate to a varchar in MySQL.
To set the properties, create a new __construct() method with a $stripeEventId argument. Inside, set that on the property and also set $this->handledAt to a new \DateTime() to set this field to "right now".