22.3.Webhooks: Preventing Replay Attacks{Preventing the Replay Attack; Update the Test}
Preventing the Replay Attack
Finally, in WebhookController, start by querying to see if this event has been handled before. Fetch the EntityManager, and then add $existingLog = $em->getRepository('App:StripeEventLog') and call findOneBy() on it to query for stripeEventId set to $eventId.
If an $existingLog is found, then we don't want to handle this. Just return a new Response() that says "Event previously handled".
If you also want to log a message so that you know when this happens, that's not a bad idea.
But if there is not an existing log, time to process this webhook! Create a new StripeEventLog and pass it $eventId. Then, persist and flush just the log.

And yea, replay attacks are gone!

Update the Test!
To make sure we didn't mess anything up, open WebhookControllerTest and copy our test method. Run that:

./bin/phpunit --filter testStripeCustomerSubscriptionDeleted
Bah! Of course... it failed for a silly reason: I need to update my test database - to add the new table. A shortcut to do that is:

./bin/console doctrine:schema:update --force --env=test
Try the test now:

./bin/phpunit --filter testStripeCustomerSubscriptionDeleted
It works! So hey, run it again!

./bin/phpunit --filter testStripeCustomerSubscriptionDeleted
It fails?!

Failed to assert that true is false.

Well, that's not clear, but I know what the problem is: every event in the test has the same event ID.
}
So when you run the test the second time, this already exists in the StripeEventLog table and the webhook is skipped. Well hey, at least we know the replay attack system is working.